Main font change
To add better compatibility to iOS devices, the font has been uploaded to the website for the main page. Other references will be changed in a future commit. Also, this is the first commit on the current account, which will do all the work from now on, with @onlinediscoveries still leading the project.